requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --refresh
annotated-types==0.5.0
    # via pydantic
anyconfig==0.13.0
    # via markata
arrow==1.2.3
    # via cookiecutter
beautifulsoup4==4.12.2
    # via markata
binaryornot==0.4.4
    # via cookiecutter
certifi==2023.7.22
    # via requests
chardet==5.2.0
    # via binaryornot
charset-normalizer==3.2.0
    # via requests
checksumdir==1.2.0
    # via markata
click==8.0.4
    # via
    #   -r requirements.in
    #   cookiecutter
    #   typer
colorama==0.4.6
    # via copier
commonmark==0.9.1
    # via markata
contourpy==1.3.1
    # via matplotlib
cookiecutter==2.3.0
    # via markata
copier==8.3.0
    # via markata
cycler==0.12.1
    # via matplotlib
dateparser==1.1.8
    # via markata
decorator==5.1.1
    # via copier
deepmerge==1.1.0
    # via markata
diskcache==5.6.3
    # via markata
dunamai==1.18.0
    # via copier
faker==19.3.1
    # via polyfactory
feedgen==0.9.0
    # via markata
fonttools==4.55.4
    # via matplotlib
funcy==2.0
    # via copier
future==0.18.3
    # via python-twitter
idna==3.4
    # via requests
jinja2==3.1.2
    # via
    #   cookiecutter
    #   copier
    #   jinja2-ansible-filters
    #   markata
    #   markata-gh
jinja2-ansible-filters==1.3.2
    # via copier
kiwisolver==1.4.8
    # via matplotlib
linkify-it-py==2.0.2
    # via markata
lxml==4.9.3
    # via feedgen
markata==0.10.0.dev4
    # via
    #   -r requirements.in
    #   markata-gh
markata-gh==0.7.0
    # via -r requirements.in
markdown==3.4.4
    # via pymdown-extensions
markdown-it-py==3.0.0
    # via
    #   markata
    #   mdit-py-plugins
    #   rich
markdown2==2.4.10
    # via markata
markupsafe==2.1.3
    # via jinja2
matplotlib==3.10.0
    # via
    #   -r requirements.in
    #   seaborn
mdit-py-plugins==0.4.0
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.1.0
    # via markata
numpy==2.2.2
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   copier
    #   dunamai
    #   matplotlib
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
pathspec==0.11.2
    # via
    #   copier
    #   markata
pillow==10.0.0
    # via
    #   markata
    #   matplotlib
pluggy==1.3.0
    # via markata
plumbum==1.8.2
    # via copier
polyfactory==2.8.0
    # via markata
prompt-toolkit==3.0.36
    # via questionary
pydantic==2.3.0
    # via
    #   copier
    #   markata
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.6.3
    # via pydantic
pydantic-extra-types==2.1.0
    # via markata
pydantic-settings==2.0.3
    # via markata
pygments==2.16.1
    # via
    #   copier
    #   markdown2
    #   rich
pyinstrument==4.5.1
    # via -r requirements.in
pymdown-extensions==10.3
    # via markata
pyparsing==3.2.1
    # via matplotlib
pyqrcode==1.2.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   dateparser
    #   faker
    #   feedgen
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via pydantic-settings
python-frontmatter==1.0.0
    # via markata
python-slugify==8.0.1
    # via
    #   cookiecutter
    #   markata
python-twitter==3.5
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   dateparser
    #   markata
    #   pandas
pyyaml==6.0.1
    # via
    #   cookiecutter
    #   copier
    #   jinja2-ansible-filters
    #   pymdown-extensions
    #   python-frontmatter
    #   pyyaml-include
    #   wavedrom
pyyaml-include==1.3.1
    # via copier
questionary==2.0.0
    # via copier
regex==2023.8.8
    # via dateparser
requests==2.31.0
    # via
    #   cookiecutter
    #   python-twitter
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via python-twitter
rich==13.5.2
    # via
    #   cookiecutter
    #   markata
seaborn==0.13.2
    # via -r requirements.in
setuptools==75.8.0
    # via
    #   anyconfig
    #   markata
six==1.16.0
    # via
    #   python-dateutil
    #   wavedrom
soupsieve==2.5
    # via beautifulsoup4
svgwrite==1.4.3
    # via wavedrom
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via markata
tomli==2.2.1
    # via markata
typer==0.9.0
    # via markata
typing-extensions==4.7.1
    # via
    #   polyfactory
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2025.1
    # via pandas
tzlocal==5.0.1
    # via dateparser
uc-micro-py==1.0.2
    # via linkify-it-py
urllib3==2.0.4
    # via requests
wavedrom==2.0.3.post3
    # via markdown2
wcwidth==0.2.6
    # via prompt-toolkit
xxhash==3.5.0
    # via markata
